backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
PyJWT==2.10.1
python-multipart==0.0.20
mutagen==1.47.0
yt-dlp==2025.2.19
pydantic-settings==2.10.1
